Skylighting.Core

Documentation

lookupSyntax :: Text -> SyntaxMap -> Maybe Syntax

Lookup syntax by (in order) full name (case insensitive), short name (case insensitive), extension.

syntaxByName :: SyntaxMap -> Text -> Maybe Syntax

Lookup a syntax by full name (case insensitive).

syntaxByShortName :: SyntaxMap -> Text -> Maybe Syntax

Lookup a syntax by short name (case insensitive).

syntaxesByExtension :: SyntaxMap -> String -> [Syntax]

Returns a list of syntaxes appropriate for the given file extension.

syntaxesByFilename :: SyntaxMap -> String -> [Syntax]

Returns a list of syntaxes appropriate for the given filename.

module Skylighting.Types

module Skylighting.Tokenizer

module Skylighting.Parser

module Skylighting.Regex

module Skylighting.Styles

module Skylighting.Loader

module Skylighting.Format.ANSI

module Skylighting.Format.HTML

module Skylighting.Format.LaTeX