Skylighting.Loader

Description

This module provides routines to load syntax definitions from disk files.

Documentation

loadSyntaxFromFile :: FilePath -> IO (Either String Syntax)

Loads a syntax definition from the specified file path. The file path must refer to a file containing an XML Kate syntax definition.

loadSyntaxesFromDir :: FilePath -> IO (Either String SyntaxMap)

Loads all syntax definitions from the specified directory by looking for files with an ".xml" extension. This function assumes such files are Kate XML syntax definitions, so XML files with unexpected contents will cause a parsing error returned as a Left.