Skylighting.Regex

Documentation

data Regex

Constructors

MatchAnyChar

MatchDynamic !Int

MatchChar (Char -> Bool)

MatchSome !Regex

MatchAlt !Regex !Regex

MatchConcat !Regex !Regex

MatchCapture !Int !Regex

MatchCaptured !Int

AssertWordBoundary

AssertBeginning

AssertEnd

AssertPositive !Direction !Regex

AssertNegative !Direction !Regex

Possessive !Regex

Lazy !Regex

Recurse

MatchNull

data RE

A representation of a regular expression.

Constructors

RE

Fields

reString :: ByteString

reCaseSensitive :: Bool

compileRegex :: Bool -> ByteString -> Either String Regex

Compile a UTF-8 encoded ByteString as a Regex. If the first parameter is True, then the Regex will be case sensitive.

matchRegex :: Regex -> ByteString -> Maybe (ByteString, IntMap (Int, Int))

Match a Regex against a (presumed UTF-8 encoded) ByteString, returning the matched text and a map of (offset, size) pairs for captures. Note that all matches are from the beginning of the string (a ^ anchor is implicit). Note also that to avoid pathological performance in certain cases, the matcher is limited to considering 2000 possible matches at a time; when that threshold is reached, it discards smaller matches. Hence certain regexes may incorrectly fail to match: e.g. a*a{3000}$ on a string of 3000 as.

testRegex :: Bool -> String -> String -> Maybe (String, [(Int, String)])

isWordChar :: Char -> Bool