Skylighting.Types

Description

Basic types for Skylighting.

Syntax descriptions

type ContextName = (Text, Text)

Full name of a context: the first member of the pair is the full syntax name, the second the context name within that syntax.

data KeywordAttr

Attributes controlling how keywords are interpreted.

Constructors

KeywordAttr

Fields

keywordCaseSensitive :: !Bool

keywordDelims :: !(Set Char)

data WordSet a

A set of "words," possibly case insensitive.

Constructors

CaseSensitiveWords !(Set a)

CaseInsensitiveWords !(Set a)

makeWordSet :: (FoldCase a, Ord a) => Bool -> [a] -> WordSet a

A set of words to match (either case-sensitive or case-insensitive).

inWordSet :: (FoldCase a, Ord a) => a -> WordSet a -> Bool

Test for membership in a WordSet.

data Matcher

Matchers correspond to the element types in a context.

Constructors

DetectChar !Char

Detect2Chars !Char !Char

AnyChar !(Set Char)

RangeDetect !Char !Char

StringDetect !Text

WordDetect !Text

RegExpr !RE

Keyword KeywordAttr !(WordSet Text)

Int

Float

HlCOct

HlCHex

HlCStringChar

HlCChar

LineContinue

IncludeRules !ContextName

DetectSpaces

DetectIdentifier

data Rule

A rule corresponds to one of the elements of a Kate syntax highlighting "context."

Constructors

Rule

Fields

rMatcher :: !Matcher

rAttribute :: !TokenType

rIncludeAttribute :: !Bool

rDynamic :: !Bool

rCaseSensitive :: !Bool

rChildren :: ![Rule]

rLookahead :: !Bool

rFirstNonspace :: !Bool

rColumn :: !(Maybe Int)

rContextSwitch :: ![ContextSwitch]

data Context

A Context corresponds to a context element in a Kate syntax description.

Constructors

Context

Fields

cName :: !Text

cSyntax :: !Text

cRules :: ![Rule]

cAttribute :: !TokenType

cLineEmptyContext :: ![ContextSwitch]

cLineEndContext :: ![ContextSwitch]

cLineBeginContext :: ![ContextSwitch]

cFallthrough :: !Bool

cFallthroughContext :: ![ContextSwitch]

cDynamic :: !Bool

data ContextSwitch

A context switch, either pops or pushes a context.

Constructors

Pop

Push !ContextName

data Syntax

A syntax corresponds to a complete Kate syntax description. The sShortname field is derived from the filename.

Constructors

Syntax

Fields

sName :: !Text

sFilename :: !String

sShortname :: !Text

sContexts :: !(Map Text Context)

sAuthor :: !Text

sVersion :: !Text

sLicense :: !Text

sExtensions :: ![String]

sStartingContext :: !Text

type SyntaxMap = Map Text Syntax

A map of syntaxes, keyed by full name.

Tokens

type Token = (TokenType, Text)

A pair consisting of a list of attributes and some text.

data TokenType

KeywordTok corresponds to dsKeyword in Kate syntax descriptions, and so on.

Constructors

KeywordTok

DataTypeTok

DecValTok

BaseNTok

FloatTok

ConstantTok

CharTok

SpecialCharTok

StringTok

VerbatimStringTok

SpecialStringTok

ImportTok

CommentTok

DocumentationTok

AnnotationTok

CommentVarTok

OtherTok

FunctionTok

VariableTok

ControlFlowTok

OperatorTok

BuiltInTok

ExtensionTok

PreprocessorTok

AttributeTok

RegionMarkerTok

InformationTok

WarningTok

AlertTok

ErrorTok

NormalTok

type SourceLine = [Token]

A line of source: a list of labeled tokens.

newtype LineNo

Line numbers

Constructors

LineNo

Fields

lineNo :: Int

Styles

data TokenStyle

A TokenStyle determines how a token is to be rendered.

Constructors

TokenStyle

Fields

tokenColor :: !(Maybe Color)

tokenBackground :: !(Maybe Color)

tokenBold :: !Bool

tokenItalic :: !Bool

tokenUnderline :: !Bool

defStyle :: TokenStyle

Default style.

data Color

A color (redgreenblue).

Constructors

RGB Word8 Word8 Word8

class ToColor a where

Things that can be converted to a color.

Methods

toColor :: a -> Maybe Color

class FromColor a where

Different representations of a Color.

Methods

fromColor :: Color -> a

data Style

A rendering style. This determines how each kind of token is to be rendered, and sets a default color and background color for normal tokens. Line numbers can have a different color and background color.

Constructors

Style

Fields

tokenStyles :: !(Map TokenType TokenStyle)

defaultColor :: !(Maybe Color)

backgroundColor :: !(Maybe Color)

lineNumberColor :: !(Maybe Color)

lineNumberBackgroundColor :: !(Maybe Color)

data ANSIColorLevel

The available levels of color complexity in ANSI terminal output.

Constructors

ANSI16Color
16-color mode

ANSI256Color
256-color mode

ANSITrueColor
True-color mode

newtype Xterm256ColorCode

Constructors

Xterm256ColorCode

Fields

getXterm256ColorCode :: Word8

Format options

data FormatOptions

Options for formatting source code.

Constructors

FormatOptions

Fields

numberLines :: !Bool

Number lines

startNumber :: !Int

Number of first line

lineAnchors :: !Bool

Anchors on each line number

titleAttributes :: !Bool

Html titles with token types

codeClasses :: ![Text]

Additional classes for Html code tag

containerClasses :: ![Text]

Additional classes for Html container tag

lineIdPrefix :: !Text

Prefix for id attributes on lines

ansiColorLevel :: !ANSIColorLevel

Level of ANSI color support to use

defaultFormatOpts :: FormatOptions

Default formatting options.